+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+## [0.7.0] - 2023-01-17
+
+### Changed
+
+- Removed Microsoft Graph specific defaults.
+- Fixed an issue where the access token provider would fail on return.
+
 ## [0.6.0] - 2023-01-16
 
 ### Changed